requirements.txt
agent==0.1.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastmcp==0.4.1
firecrawl-py==1.15.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
griffe==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
motor==3.7.0
multidict==6.4.3
narwhals==1.33.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.72.0
openai-agents==0.0.7
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
playwright==1.51.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.19.0
youtube-transcript-api==1.0.3
